requirements.txt
altair==4.2.2
attrs==22.2.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
blinker==1.6.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
entrypoints==0.4
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
protobuf==3.20.3
psycopg2==2.9.6
pyarrow==11.0.0
pycodestyle==2.10.0
pydeck==0.8.0
Pygments==2.15.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
rich==13.3.3
six==1.16.0
smmap==5.0.0
streamlit==1.21.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0
